requirements.txt
aniso8601==1.1.0
Flask==0.10.1
Flask-RESTful==0.3.4
flask-restplus==0.8.0
functools32==3.2.3.post2
itsdangerous==0.24
Jinja2==2.8
jsonschema==2.5.1
MarkupSafe==0.23
mysql-connector-python==2.0.4
numpy==1.10.1
pandas==0.17.0
python-dateutil==2.4.2
pytz==2015.7
six==1.10.0
Werkzeug==0.10.4
wheel==0.24.0
